#111 Alterando nome do arquivo e realizando requisição
